Add basic database model
Adds basic database model to get us started. This will need to be
extended drastically for basic functionality but should allow
us to make a simple worker that simply adds hashes to the DB.

Long term maybe MySQL isnt even a good idea, but it'll get a
basic proof of concept sorted for now.

Change-Id: Ifdcd2309fa2ad6a9a27dd6c3ba61dd8db0dcb854